refactor: partner contact page
