+++ TrashMob/client-app/src/App.tsx
@@ -43,14 +43,16 @@ import { useLogin } from './hooks/useLogin';
 import { Home as Home2024 } from './pages/Home';
 import { TermsOfService } from './pages/termsofservice';
 import { VolunteerOpportunities } from './pages/volunteeropportunities';
-import { PartnerEdit } from './pages/partnerdashboard/[partnerId]/edit';
-import { PartnerLayout } from './pages/partnerdashboard/[partnerId]/_layout';
+import { PartnerEdit } from './pages/partnerdashboard/$partnerId/edit';
+import { PartnerLayout } from './pages/partnerdashboard/$partnerId/_layout';
 import { PartnerLocations } from './components/Partners/PartnerLocations';
-import { PartnerContacts } from './pages/partnerdashboard/[partnerId]/contacts';
+import { PartnerContacts } from './pages/partnerdashboard/$partnerId/contacts';
 import { PartnerAdmins } from './components/Partners/PartnerAdmins';
 import { PartnerDocuments } from './components/Partners/PartnerDocuments';
 import { PartnerSocialMediaAccounts } from './components/Partners/PartnerSocialMediaAccounts';
-import { PartnerIndex } from './pages/partnerdashboard/[partnerId]';
+import { PartnerIndex } from './pages/partnerdashboard/$partnerId';
+import { PartnerContactEdit } from './pages/partnerdashboard/$partnerId/contacts.$contactId.edit';
+import { PartnerContactCreate } from './pages/partnerdashboard/$partnerId/contacts.create';
 
 const queryClient = new QueryClient();
 
@@ -127,7 +129,10 @@ export const App: FC = () => {
                                             <Route index element={<PartnerIndex />} />
                                             <Route path='edit' element={<PartnerEdit />} />
                                             <Route path='locations' element={<PartnerLocations />} />
-                                            <Route path='contacts' element={<PartnerContacts />} />
+                                            <Route path='contacts' element={<PartnerContacts />}>
+                                                <Route path='create' element={<PartnerContactCreate />} />
+                                                <Route path=':contactId/edit' element={<PartnerContactEdit />} />
+                                            </Route>
                                             <Route path='admins' element={<PartnerAdmins />} />
                                             <Route path='documents' element={<PartnerDocuments />} />
                                             <Route path='socials' element={<PartnerSocialMediaAccounts />} />
