+++ TrashMob/client-app/src/pages/partnerdashboard/$partnerId/index.tsx
